--- Project03/StoreFront/ProductCellTableViewCell.swift
@@ -1,23 +1,0 @@
-//
-//  ProductCellTableViewCell.swift
-//  Project03
-//
-//  Created by Lea W. Leonard on 5/23/21.
-//
-
-import UIKit
-
-class ProductCellTableViewCell: UITableViewCell {
-
-    override func awakeFromNib() {
-        super.awakeFromNib()
-        // Initialization code
-    }
-
-    override func setSelected(_ selected: Bool, animated: Bool) {
-        super.setSelected(selected, animated: animated)
-
-        // Configure the view for the selected state
-    }
-
-}
